Implement Opus packet frame parsing
Implement Opus packet frame parsing for codes 0, 1, 2, and 3.
Enforce RFC 6716 malformed packet rules R1-R7, fix TOC
config/frame-duration parsing bugs used by packet validation, and
add packet parser tests covering malformed rules and zero-length
edge cases.

Validation:
- go test ./... -count=1
- golangci-lint run

+++ internal/bitdepth/bitdepth.go
@@ -18,7 +18,7 @@ func ConvertFloat32LittleEndianToSigned16LittleEndian(in []float32, out []byte, 
 			out[currIndex] = byte(res & 0b11111111)
 			currIndex++
 
-			out[currIndex] = (byte(res >> 8))
+			out[currIndex] = byte(uint16(res) >> 8) // #nosec G115
 			currIndex++
 		}
 	}
